Use more generic check_memory name
