--- cmd/check_available_memory/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/ncr-devops-platform/nagiosfoundation/cmd/check_available_memory/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
