+++ cmd/check_memory/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/ncr-devops-platform/nagiosfoundation/cmd/check_memory/cmd"
+)
+
+func main() {
+	cmd.Execute()
+}
