+++ lib/app/nagiosfoundation/check_memory.go
@@ -8,12 +8,12 @@ import (
 	"github.com/ncr-devops-platform/nagiosfoundation/lib/pkg/nagiosformatters"
 )
 
-// CheckAvailableMemoryWithHandler determines the percentage of
+// CheckMemoryWithHandler determines the percentage of
 // memory used and emits a critical response if it's over
 // the critical argument, a warning response if it's over the
 // warning argument, and good response otherwise.
-func CheckAvailableMemoryWithHandler(warning, critical int, metricName string, memoryHandler func() uint64) (string, int) {
-	const checkName = "CheckAvailableMemory"
+func CheckMemoryWithHandler(checkType string, warning, critical int, metricName string, memoryHandler func() uint64) (string, int) {
+	const checkName = "CheckMemory"
 
 	var msg string
 	var retcode int
@@ -36,11 +36,11 @@ func CheckAvailableMemoryWithHandler(warning, critical int, metricName string, m
 	return msg, retcode
 }
 
-// CheckAvailableMemory executes CheckAvailableMemoryWithHandler(),
+// CheckMemory executes CheckMemoryWithHandler(),
 // passing it the OS constranted GetFreeMemory() function, prints
 // the returned message and exits with the returned exit code.
 //
-// Returns are those of CheckAvailableMemoryWithHandler()
-func CheckAvailableMemory(warning, critical int, metricName string) (string, int) {
-	return CheckAvailableMemoryWithHandler(warning, critical, metricName, memory.GetUsedMemoryPercentage)
+// Returns are those of CheckMemoryWithHandler()
+func CheckMemory(checkType string, warning, critical int, metricName string) (string, int) {
+	return CheckMemoryWithHandler(checkType, warning, critical, metricName, memory.GetUsedMemoryPercentage)
 }
